main.c divisiob float

+++ division.c
@@ -1,4 +1,4 @@
-int division(int a, int b)
+float division(int a, int b)
 {
     return a / b;
